Refactored satellite simulation by adding new satellite classes (Sputnik, CrewDragon, Starlink) and updating GPS parameters. Enhanced project files to include new source and header files, and adjusted GPS initialization in the simulation. Updated test cases to reflect changes in GPS positioning and radius.
