+++ W10/231.09.Lab/crewDragon.cpp
@@ -1,0 +1,16 @@
+/***********************************************************************
+ * crewDragon.cpp
+ * Crew Dragon implementation: LEO per program spec.
+ ************************************************************************/
+
+#include "crewDragon.h"
+
+CrewDragon::CrewDragon() : Satellite(0.0, 8000000.0, -7900.0, 0.0)
+{
+   radius = 7.0;
+}
+
+void CrewDragon::draw(ogstream& gout)
+{
+   gout.drawCrewDragon(position, direction.getRadians());
+}
