+++ W10/231.09.Lab/crewDragon.h
@@ -1,0 +1,15 @@
+/***********************************************************************
+ * crewDragon.h
+ * Crew Dragon: low Earth orbit, radius 7 pixels.
+ ************************************************************************/
+
+#pragma once
+
+#include "satellite.h"
+
+class CrewDragon : public Satellite
+{
+public:
+   CrewDragon();
+   void draw(ogstream& gout) override;
+};
