+++ W10/231.09.Lab/gps.h
@@ -1,0 +1,16 @@
+/***********************************************************************
+ * gps.h
+ * GPS satellite: navigation orbit, radius 12 pixels. Six orbits supported.
+ ************************************************************************/
+
+#pragma once
+
+#include "satellite.h"
+
+class GPS : public Satellite
+{
+public:
+   GPS();
+   GPS(double x, double y, double vx, double vy);
+   void draw(ogstream& gout) override;
+};
