+++ W10/231.09.Lab/hubble.cpp
@@ -1,0 +1,16 @@
+/***********************************************************************
+ * hubble.cpp
+ * Hubble implementation: GEO orbit per program spec.
+ ************************************************************************/
+
+#include "hubble.h"
+
+Hubble::Hubble() : Satellite(0.0, -GEO_DISTANCE, GEO_VELOCITY, 0.0)
+{
+   radius = 10.0;
+}
+
+void Hubble::draw(ogstream& gout)
+{
+   gout.drawHubble(position, direction.getRadians());
+}
