+++ W10/231.09.Lab/starlink.h
@@ -1,0 +1,15 @@
+/***********************************************************************
+ * starlink.h
+ * Starlink satellite: low Earth orbit, radius 6 pixels.
+ ************************************************************************/
+
+#pragma once
+
+#include "satellite.h"
+
+class Starlink : public Satellite
+{
+public:
+   Starlink();
+   void draw(ogstream& gout) override;
+};
